Add ReImage script for WS 2019
Added a couple of Output statements to be able to follow the process along and updated the script for Server 2019.  Easy stuff but I had to change the .NET 3.5 handling due to a different CAB naming scheme.  Using Add-WindowsCapability avoids the naming problem and is base on a blog post by Neihaus.